PyTorch implementation of the networks.
TODO: CWT Target Network, Spectrogram Source Network, Spectrogram Target Network. Note that the current implementation employ the PReLU activation function and as such perform worst than the PELU version in Theano. This should be fixed soon. Also, the number of parameters between the two implementation are differents due to PyTorch not counting the running mean and variance as parameters of the model while Lasagne does.